back-end/requirements.txt
beautifulsoup4==4.11.2
char_featurizer==1.0.0
fire==0.4.0
fonttools==4.37.1
fonttools==4.25.0
lpips==0.1.4
matplotlib==3.5.2
numpy==1.21.6
numpy==1.21.5
Pillow==9.3.0
Pillow==9.4.0
Pillow==9.5.0
Requests==2.29.0
scikit_image==0.19.3
scipy==1.7.3
sconf==0.2.5
setuptools==65.6.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.63.1
